+++ src/TorchSharp/Tensor/Tensor.cs
@@ -1415,12 +1415,6 @@ namespace TorchSharp
                 set { index_put_(value, indices); }
             }
 
-            [IndexerName("TensorItems")]
-            public Tensor this[params Tensor[] indices] {
-                get { return index(indices); }
-                set { index_put_(value, indices); }
-            }
-
             /// <summary>
             /// Tensor indexer.
             /// </summary>
@@ -1572,14 +1566,6 @@ namespace TorchSharp
                         }
                     }
                 }
-            }
-
-            /// <summary>
-            /// Index into the tensor using Python-like indexing expressions.
-            /// </summary>
-            public Tensor index(params Tensor[] indices)
-            {
-                return index(indices.Select(t => TensorIndex.Tensor(t)).ToArray());
             }
 
             /// <summary>
@@ -7143,18 +7129,15 @@ namespace TorchSharp
                 return TensorIndex.Single(value);
             }
 
-            public static implicit operator Tensor(TensorIndex value)
+            public static implicit operator TensorIndex(Tensor tensor)
             {
-                _throw();
-                return new Tensor(IntPtr.Zero);
+                return TensorIndex.Tensor(tensor);
             }
 
-            private static void _throw()
+            public static implicit operator TensorIndex((int? start, int? end) range)
             {
-                throw new InvalidOperationException("Should not be called.");
+                return TensorIndex.Slice(range.start, range.end);
             }
-
-            public static implicit operator TensorIndex((int? start, int? end) range) => TensorIndex.Slice((long?)range.start, (long?)range.end);
 
 #if !NETSTANDARD2_0_OR_GREATER
             public static implicit operator TensorIndex(System.Range range)