+++ RELEASENOTES.md
@@ -8,6 +8,10 @@ __Breaking Changes__:
 
 __API Changes__:
 
+- #1266 Implicit operator from `Tensor` to `TensorIndex` has been added.
+    - The implicit operator from `TensorIndex` to `Tensor` has been removed.
+    - The indexer that accepts tensors as the indices has been removed.
+
 __Bug Fixes__: